refactor: Move set, file, and array function tests to dedicated files
Migrate 127 tests (48 set + 39 file + 40 array) to improve separation
of test code from implementation
