+++ src/vm/tests/mod.rs
@@ -3,3 +3,6 @@ mod basic;
 mod math_variadic;
 mod math_errors;
 mod string_functions;
+mod array_functions;
+mod file_functions;
+mod set_functions;
